refactor: 모임방 참여자 관련 클래스 디렉토리 변경
